End April update
Continue refactoring code.

+++ clkernels/intp_plus_version2.cl
@@ -1,16 +1,13 @@
-__kernel void intp_plus_version2(__global const float *blend, __global const float *tmpI, __global float *lum, __global const int *X, __global const int *Y, const double amount) {
+__kernel void intp_plus_version2(__global const float *blend, __global const float *tmpI, __global float *lum, __global const int *X, __global const int *Y, const float amount) {
  
     // Get the index of the current element to be processed
     int index = get_global_id(0);
     int currentY = Y[index]; int currentX = X[index];
     
     //from rt_math
-    // calculate a * b + (1 - a) * c
-    // following is valid:
-    // intp(a, b+x, c+x) = intp(a, b, c) + x
-    // intp(a, b*x, c*x) = intp(a, b, c) * x
+    // return a * (b - c) + c;
 
-    /*luminance[i][j] = intp(blend_jagged_array[i][j] * amount, max(tmpI[i][j], 0.0f), luminance[i][j]);*/
+    //*luminance[i][j] = intp(blend_jagged_array[i][j] * amount, max(tmpI[i][j], 0.0f), luminance[i][j]);*/
 
     double temp_term = blend[index] * amount;
